+++ samples/kotlin-maven-simple/src/main/kotlin/org/refactorkit/samples/Greeting.kt
@@ -1,3 +1,6 @@
 package org.refactorkit.samples
 
 class Greeting
+interface GreetingPort
+enum class GreetingMode { HELLO }
+annotation class GreetingMarker